+++ app/Http/Controllers/ConsortiaMembersController.php
@@ -11,40 +11,47 @@ class ConsortiaMembersController extends Controller
     public function addConsortiaMember(Request $request){
         $this->validate($request, array(
             'acronym' => 'required|max:255',
-            'name' => 'required'
+            'name' => 'required',
+            'image' => ['mimes:jpeg,jpg,png,gif', 'max:10240', 'required'],
+            'consortia' => 'required'
         ));
-
         $user = auth()->user();
-        $consortia_member = new ConsortiaMember;
-        $consortia_member->acronym = $request->acronym;
-        $consortia_member->name = $request->name;
-        $consortia_member->profile = $request->profile;
-        $consortia_member->contact_name = $request->contact_name;
-        $consortia_member->contact_details = $request->contact_details;
-        if($request->website){
-            if (!preg_match("~^(?:f|ht)tps?://~i", $request->website)) {
-                $consortia_member->website = "http://" . $request->website;
-            }
-        }
-        
-        $consortia_member->consortia_id = $request->consortia;
-        if($request->hasFile('image')){
-            $imageFile = $request->file('image');
-            $imageName = uniqid().$imageFile->getClientOriginalName();
-            $imageFile->move(public_path('/storage/page_images/'), $imageName);
-            $consortia_member->thumbnail = $imageName;
-        }
-        $consortia_member->save();
-
+        $temp_changes = '';
         $log = new Log;
-        $log->user_id = $user->id;
-        $log->user_email = $user->email;
-        $log->action = 'Added \''. $consortia_member->acronym.'\' to consortia members';
-        $log->IP_address = $request->ip();
-        $log->resource = 'Consortia Member';
-        $log->save();
+        if($user->role != 5 && $user->role != 1){
+            return redirect()->back()->with('error','Your account is not authorized to use this function.'); 
+        } else {
+            $consortia_member = new ConsortiaMember;
+            $consortia_member->acronym = $request->acronym;
+            $consortia_member->name = $request->name;
+            $consortia_member->profile = $request->profile;
+            $consortia_member->contact_name = $request->contact_name;
+            $consortia_member->contact_details = $request->contact_details;
+            if($request->website){
+                if (!preg_match("~^(?:f|ht)tps?://~i", $request->website)) {
+                    $consortia_member->website = "http://" . $request->website;
+                }
+            }
+            
+            $consortia_member->consortia_id = $request->consortia;
+            if($request->hasFile('image')){
+                $imageFile = $request->file('image');
+                $imageName = uniqid().$imageFile->getClientOriginalName();
+                $imageFile->move(public_path('/storage/page_images/'), $imageName);
+                $consortia_member->thumbnail = $imageName;
+            }
+            $consortia_member->save();
 
-        return redirect()->back()->with('success','Consortia Member Added.'); 
+            $log->user_id = $user->id;
+            $log->user_email = $user->email;
+            $log->changes = '<strong>Added: </strong>\''. $consortia_member->acronym.'\' to consortia members';
+            $log->action = 'Added \''. $consortia_member->acronym.'\' to consortia members';
+            $log->IP_address = $request->ip();
+            $log->resource = 'Consortia Member';
+            $log->save();
+
+            return redirect()->back()->with('success','Consortia Member Added.'); 
+        }
     }
     
     public function editConsortiaMember(Request $request, $consortia_member_id){
@@ -52,45 +59,73 @@ class ConsortiaMembersController extends Controller
             'acronym' => 'required|max:255',
             'name' => 'required'
         ));
-        
         $user = auth()->user();
-        $consortia_member = ConsortiaMember::find($consortia_member_id);
-        $consortia_member->acronym = $request->acronym;
-        $consortia_member->name = $request->name;
-        $consortia_member->profile = $request->profile;
-        $consortia_member->contact_name = $request->contact_name;
-        $consortia_member->contact_details = $request->contact_details;
-        if($request->website){
-            if (!preg_match("~^(?:f|ht)tps?://~i", $request->website)) {
-                $consortia_member->website = "http://" . $request->website;
-            } else {
-                $consortia_member->website = $request->website;
+        $temp_changes = '';
+        $log = new Log;
+        if($user->role != 5 && $user->role != 1){
+            return redirect()->back()->with('error','Your account is not authorized to use this function.'); 
+        } else {
+            $consortia_member = ConsortiaMember::find($consortia_member_id);
+
+            if($consortia_member->acronym != $request->acronym){
+                $temp_changes = $temp_changes.'<strong>Acronym:</strong> '.$consortia_member->acronym.' <strong>-></strong> '.$request->acronym.'<br>';
             }
-        }
-        $consortia_member->consortia_id = $request->consortia;
-        if($request->hasFile('image')){
-            if($consortia_member->thumbnail != null){
-                $image_path = public_path().'/storage/page_images/'.$consortia_member->thumbnail;
-                if(file_exists($image_path)){
-                    unlink($image_path);
+            if($consortia_member->name != $request->name){
+                $temp_changes = $temp_changes.'<strong>Name:</strong> '.$consortia_member->name.' <strong>-></strong> '.$request->name.'<br>';
+            }
+            if($consortia_member->website != $request->website){
+                $temp_changes = $temp_changes.'<strong>Website:</strong> '.$consortia_member->website.' <strong>-></strong> '.$request->website.'<br>';
+            }
+            if($consortia_member->profile != $request->profile){
+                $temp_changes = $temp_changes.'<strong>Profile:</strong> '.$consortia_member->profile.' <strong>-></strong> '.$request->profile.'<br>';
+            }
+            if($consortia_member->consortia_id != $request->consortia){
+                $temp_changes = $temp_changes.'<strong>Consortia ID:</strong> '.$consortia_member->consortia_id.' <strong>-></strong> '.$request->consortia.'<br>';
+            }
+            if($consortia_member->contact_name != $request->contact_name){
+                $temp_changes = $temp_changes.'<strong>Contact Name:</strong> '.$consortia_member->contact_name.' <strong>-></strong> '.$request->contact_name.'<br>';
+            }
+            if($consortia_member->contact_details != $request->contact_details){
+                $temp_changes = $temp_changes.'<strong>Contact Image:</strong> '.$consortia_member->contact_details.' <strong>-></strong> '.$request->contact_details.'<br>';
+            }
+
+            $consortia_member->acronym = $request->acronym;
+            $consortia_member->name = $request->name;
+            $consortia_member->profile = $request->profile;
+            $consortia_member->contact_name = $request->contact_name;
+            $consortia_member->contact_details = $request->contact_details;
+            if($request->website){
+                if (!preg_match("~^(?:f|ht)tps?://~i", $request->website)) {
+                    $consortia_member->website = "http://" . $request->website;
+                } else {
+                    $consortia_member->website = $request->website;
                 }
             }
-            $imageFile = $request->file('image');
-            $imageName = uniqid().$imageFile->getClientOriginalName();
-            $imageFile->move(public_path('/storage/page_images/'), $imageName);
-            $consortia_member->thumbnail = $imageName;
+            $consortia_member->consortia_id = $request->consortia;
+            if($request->hasFile('image')){
+                if($consortia_member->thumbnail != null){
+                    $image_path = public_path().'/storage/page_images/'.$consortia_member->thumbnail;
+                    if(file_exists($image_path)){
+                        unlink($image_path);
+                    }
+                }
+                $imageFile = $request->file('image');
+                $imageName = uniqid().$imageFile->getClientOriginalName();
+                $imageFile->move(public_path('/storage/page_images/'), $imageName);
+                $consortia_member->thumbnail = $imageName;
+            }
+            $consortia_member->save();
+
+            $log->user_id = $user->id;
+            $log->changes = $temp_changes;
+            $log->user_email = $user->email;
+            $log->action = 'Edited \''. $consortia_member->acronym.'\' details';
+            $log->IP_address = $request->ip();
+            $log->resource = 'Consortia';
+            $log->save();
+
+            return redirect()->back()->with('success','Consortia Member Updated.'); 
         }
-        $consortia_member->save();
-
-        $log = new Log;
-        $log->user_id = $user->id;
-        $log->user_email = $user->email;
-        $log->action = 'Edited \''. $consortia_member->acronym.'\' details';
-        $log->IP_address = $request->ip();
-        $log->resource = 'Consortia Member';
-        $log->save();
-
-        return redirect()->back()->with('success','Consortia Member Updated.'); 
     }
 
     public function editConsortiaMemberBanner(Request $request, $consortia_member_id){
@@ -142,33 +177,67 @@ class ConsortiaMembersController extends Controller
 
     public function editConsortiaMemberDetails(Request $request, $consortia_member_id){
         $user = auth()->user();
-        $consortia_member = ConsortiaMember::find($consortia_member_id);
-        $consortia_member->profile = $request->profile;
-        $consortia_member->contact_name = $request->contact_name;
-        $consortia_member->contact_details = $request->contact_details;
-        $consortia_member->save();
-
+        $temp_changes = '';
         $log = new Log;
-        $log->user_id = $user->id;
-        $log->user_email = $user->email;
-        $log->action = 'Edited \''. $consortia_member->acronym.'\' details';
-        $log->IP_address = $request->ip();
-        $log->resource = 'Consortia Member';
-        $log->save();
+        if($user->role != 5 && $user->role != 1){
+            return redirect()->back()->with('error','Your account is not authorized to use this function.'); 
+        } else {
+            $consortia_member = ConsortiaMember::find($consortia_member_id);
 
-        return redirect()->back()->with('success','Consortia Member Details Updated.'); 
+            if($consortia_member->profile != $request->profile){
+                $temp_changes = $temp_changes.'<strong>Profile:</strong> '.$consortia_member->profile.' <strong>-></strong> '.$request->profile.'<br>';
+            }
+            if($consortia_member->contact_name != $request->contact_name){
+                $temp_changes = $temp_changes.'<strong>Contact Name:</strong> '.$consortia_member->contact_name.' <strong>-></strong> '.$request->contact_name.'<br>';
+            }
+            if($consortia_member->contact_details != $request->contact_details){
+                $temp_changes = $temp_changes.'<strong>Contact Image:</strong> '.$consortia_member->contact_details.' <strong>-></strong> '.$request->contact_details.'<br>';
+            }
+
+            $consortia_member->profile = $request->profile;
+            $consortia_member->contact_name = $request->contact_name;
+            $consortia_member->contact_details = $request->contact_details;
+            $consortia_member->save();
+
+            $log->user_id = $user->id;
+            $log->user_email = $user->email;
+            $log->changes = $temp_changes;
+            $log->action = 'Edited \''. $consortia_member->acronym.'\' details';
+            $log->IP_address = $request->ip();
+            $log->resource = 'Consortia Member';
+            $log->save();
+
+            return redirect()->back()->with('success','Consortia Member Details Updated.'); 
+        }
     }
 
-    public function deleteConsortiaMember($consortia_member_id){
-        $consortia_member = ConsortiaMember::find($consortia_member_id);
-        if($consortia_member->thumbnail != null){
-            $image_path = public_path().'/storage/page_images/'.$consortia_member->thumbnail;
-            if(file_exists($image_path)){
-                unlink($image_path);
+    public function deleteConsortiaMember(Request $request, $consortia_member_id){
+        $user = auth()->user();
+        $temp_changes = '';
+        $log = new Log;
+        if($user->role != 5 && $user->role != 1){
+            return redirect()->back()->with('error','Your account is not authorized to use this function.'); 
+        } else {
+            $consortia_member = ConsortiaMember::find($consortia_member_id);
+            if($consortia_member->thumbnail != null){
+                $image_path = public_path().'/storage/page_images/'.$consortia_member->thumbnail;
+                if(file_exists($image_path)){
+                    unlink($image_path);
+                }
             }
+            
+            $log->user_id = $user->id;
+            $log->user_email = $user->email;
+            $log->changes = '<strong>Deleted:</strong> '.$consortia_member->acronym.'';
+            $log->action = 'Deleted \''. $consortia_member->acronym.'\'';
+            $log->IP_address = $request->ip();
+            $log->resource = 'Consortia Member';
+            $log->save();
+
+            $consortia_member->delete();
+
+            return redirect()->back()->with('success','Consortia Member Deleted.'); 
         }
-        $consortia_member->delete();
-        return redirect()->back()->with('success','Consortia Member Deleted.'); 
     }
