+++ app/Http/Controllers/IndustriesController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Industry;
+use App\Log;
 
 class IndustriesController extends Controller
 {
@@ -11,35 +12,81 @@ class IndustriesController extends Controller
         $this->validate($request, array(
             'name' => 'required|max:255'
         ));
-
         $user = auth()->user();
-        $industry = new Industry;
-        $industry->name = $request->name;
-        $industry->save();
+        $temp_changes = '';
+        $log = new Log;
+        if($user->role != 5 && $user->role != 1){
+            return redirect()->back()->with('error','Your account is not authorized to use this function.'); 
+        } else {
+            $industry = new Industry;
+            $industry->name = $request->name;
+            $industry->save();
 
-        return redirect()->back()->with('success','Industry Added.'); 
+            $log->user_id = $user->id;
+            $log->user_email = $user->email;
+            $log->changes = '<strong>Added:</strong> '.$industry->name.'';
+            $log->action = 'Added \''. $industry->name.'\'';
+            $log->IP_address = $request->ip();
+            $log->resource = 'Industries';
+            $log->save();
+
+            return redirect()->back()->with('success','Industry Added.'); 
+        }
     }
     
     public function editIndustry(Request $request, $industry_id){
         $this->validate($request, array(
             'name' => 'required|max:255'
         ));
-        
         $user = auth()->user();
-        $industry = Industry::find($industry_id);
-        $industry->name = $request->name;
-        $industry->save();
+        $temp_changes = '';
+        $log = new Log;
+        if($user->role != 5 && $user->role != 1){
+            return redirect()->back()->with('error','Your account is not authorized to use this function.'); 
+        } else {
+            $industry = Industry::find($industry_id);
 
-        return redirect()->back()->with('success','Industry Updated.'); 
+            if($industry->name != $request->name){
+                $temp_changes = $temp_changes.'<strong>Name:</strong> '.$industry->name.' <strong>-></strong> '.$request->name.'<br>';
+            }
+
+            $industry->name = $request->name;
+            $industry->save();
+
+            $log->user_id = $user->id;
+            $log->user_email = $user->email;
+            $log->changes = $temp_changes;
+            $log->action = 'Edited \''. $industry->name.'\'';
+            $log->IP_address = $request->ip();
+            $log->resource = 'Industries';
+            $log->save();
+
+            return redirect()->back()->with('success','Industry Updated.'); 
+        }
     }
 
     public function deleteIndustry($industry_id, Request $request){
-        $industry = Industry::find($industry_id);
-        $deletedName = $industry->name;
-        $industry->articles()->delete();
-        $industry->delete();
+        $user = auth()->user();
+        $temp_changes = '';
+        $log = new Log;
+        if($user->role != 5 && $user->role != 1){
+            return redirect()->back()->with('error','Your account is not authorized to use this function.'); 
+        } else {
+            $industry = Industry::find($industry_id);
+            $deletedName = $industry->name;
+            $industry->articles()->delete();
+            $industry->delete();
 
-        return redirect()->back()->with('success','Industry Deleted.'); 
+            $log->user_id = $user->id;
+            $log->user_email = $user->email;
+            $log->changes = '<strong>Deleted:</strong> '.$deletedName.'';
+            $log->action = 'Deleted \''. $deletedName.'\'';
+            $log->IP_address = $request->ip();
+            $log->resource = 'Industries';
+            $log->save();
+
+            return redirect()->back()->with('success','Industry Deleted.'); 
+        }
     }
 
 }
